split tests by builder
