+++ tests/chunk_tests.rs
@@ -53,25 +53,6 @@ fn text_element() {
 }
 
 #[test]
-fn client_text_element() {
-    let template = tmpl(
-        "
-
-            Beasts tread
-            softly underfoot.
-            
-		",
-        [],
-    );
-    let expected = "Beasts tread softly underfoot.";
-
-    let mut html = ClientHtml::new();
-    let results = html.build(&template);
-
-    assert_eq!(Ok(expected.to_string()), results);
-}
-
-#[test]
 fn element_with_text() {
     let template = tmpl(
         "
